01/30/2017 3:12AM
+ Added checkpoint assets into the game. Wrote scripts to trigger when
the player has reached one. Created a ScriptableObject to keep track of
the player's location and which checkpoint they hit. If one checkpoint
is deployed, all others will go back to idle to indicate this is not the
checkpoint they will respawn at.

* Fixed the animation bug for the swarmers. Swarmers now spawn in the
jump position and transistion to walking.

+ In spriter pro, with the skeletal structure implemented by Hao, 45 new
animations were created for the main character. This includes, idle,
falling, aiming in all directions, running and aiming in all directions,
falling and aiming in all directions, and running backwards and aiming
in all directions.

* Completely redone player animation script. Nothing from the old system
could be saved and a clean slate was needed.

+ Created new scripts which aids in bullet instantiation for the new
character model.

- Deleted several scripts that which were vital to the previous way we
handled animation for the character. Now obsolete.

* Other various small changes and tweaks.

NOTE: There are so many changes in this commit because I renamed a
couple of folder for clarification. All files within these folders are
reported as changes. Sorry!  :-(

+++ Capstone Project/Assets/Scripts/Camera/Screen Shakes/CameraShakeManager.cs
@@ -10,7 +10,7 @@ public class CameraShakeManager : MonoBehaviour {
     private Queue<ScreenShakeRequest> _shakeQ = new Queue<ScreenShakeRequest>();
 
     private ControllableObject _controller;
-    private byte _key;
+    private int _key;
     private Vector3 _defaultPosition = new Vector3(0.0f, 0.0f, 0.0f);
 
     private void OnEnable() {
